move fallback handler to wrapper.py
