+++ datadog_lambda/handler.py
@@ -9,34 +9,12 @@ from importlib import import_module
 import os
 import time
 
-from datadog_lambda.tracing import emit_telemetry_on_exception_outside_of_handler
-from datadog_lambda.wrapper import datadog_lambda_wrapper
+from datadog_lambda.wrapper import datadog_lambda_wrapper, error_fallback_handler
 from datadog_lambda.module_name import modify_module_name
 
 
 class HandlerError(Exception):
     pass
-
-
-class _ErrorFallbackHandler(object):
-    """
-    Decorator for when an exception occurs outside of the handler function.
-    Emits telemetry and re-raises the exception.
-    """
-
-    def __init__(self, exception, modified_mod_name, start_time_ns):
-        self.exception = exception
-        self.modified_mod_name = modified_mod_name
-        self.start_time_ns = start_time_ns
-
-    def __call__(self, event, context, **kwargs):
-        emit_telemetry_on_exception_outside_of_handler(
-            context,
-            self.exception,
-            self.modified_mod_name,
-            self.start_time_ns,
-        )
-        raise self.exception
 
 
 path = os.environ.get("DD_LAMBDA_HANDLER", None)
@@ -58,4 +36,4 @@ try:
     handler_func = getattr(handler_module, handler_name)
     handler = datadog_lambda_wrapper(handler_func)
 except Exception as e:
-    handler = _ErrorFallbackHandler(e, modified_mod_name, start_time_ns)
+    handler = error_fallback_handler(e, modified_mod_name, start_time_ns)
